Some changes in tx.js
Moved the fetchTxHeaders function within the fastify.get endpoint so that it is only called when needed. 
This will improve performance by avoiding unnecessary network calls. 

Added error handling to the fetchTxHeaders function. 
If the GraphQL query fails or returns an error, it will return null instead of throwing an error. 

Renamed txFromWarp to contractTxFromWarp to make it clearer that it is a contract transaction object. 

Added error handling for the fetch call to gateway.warp.cc. 
If the call fails or returns a 404 status code, it will return an error object with a 404 status code. 

Replaced the use of JSON.stringify and JSON.parse with the global JSON object.