Changes in root.js
Added try/catch to handle any errors that may occur during the request/response cycle. 

Added error handling to ensure that if the gateway returns a status other than 200 or 202, we throw an error and return a response with status code 500 and an error message. 

Optimised the if statement condition by using the includes method to check that the response status is included in the array of valid statuses.